add forge_smap

+++ bootstrap.sh
@@ -3,12 +3,15 @@
 set -xe
 
 if command -v wget >/dev/null 2>&1; then
-    wget https://sqlite.org/2025/sqlite-autoconf-3500100.tar.gz
-    mv sqlite-autoconf-3500100.tar.gz ./src
-    cd src
-    tar -vxzf ./sqlite-autoconf-3500100.tar.gz
-    cd ..
+    if [ ! -e "src/sqlite-autoconf-3500100" ]; then
+        wget https://sqlite.org/2025/sqlite-autoconf-3500100.tar.gz
+        mv sqlite-autoconf-3500100.tar.gz ./src
+        cd src
+        tar -vxzf ./sqlite-autoconf-3500100.tar.gz
+        cd ..
+    fi
     autoreconf --install
+    ./configure --prefix=/usr --libdir=/usr/lib64
 else
     unset -x
     echo "wget is not installed, it is required to get sqlite."
